refactor: Moves first attempt to old/ directory
